+++ client/lib/ok.dart
@@ -4,3 +4,5 @@ export 'package:object_key/object_key.dart';
 
 export 'package:lnc/log.dart';
 export 'package:lnc/notification.dart';
+export 'package:lnc/cache.dart';
+export 'package:lnc/time.dart';